+++ app/src/main/java/com/example/movie/adapter/LikeMoviesAdapter.kt
@@ -9,7 +9,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.example.movie.DetailActivity
+import com.example.movie.view.DetailActivity
 import com.example.movie.R
 import com.example.movie.model.Movie
