+++ app/src/main/java/com/example/movie/api/RetrofitService.kt
@@ -1,6 +1,5 @@
 package com.example.movie.api
 
-import com.example.movie.model.Genre
 import com.example.movie.model.MovieResponse
 import com.google.gson.JsonObject
 import retrofit2.Call
@@ -105,7 +104,4 @@ interface PostApi {
 
     @DELETE("authentication/session")
     suspend fun deleteSessionCoroutine(@Query("api_key") apiKey: String, @Body body: JsonObject): Response<JsonObject>
-
-    @GET("genre/movie/list")
-    fun getGenres(@Query("api_key") apiKey: String): Call<List<Genre>>
 }
