--- app/src/main/java/com/example/movie/model/Genre.kt
@@ -1,10 +1,0 @@
-package com.example.movie.model
-
-import com.google.gson.annotations.SerializedName
-
-data class Genre (
-    @SerializedName("id")
-    var id: Int = 0,
-    @SerializedName("name")
-    var name: String = " "
-)
