+++ app/src/main/java/com/example/movie/myFragments/MainFragment.kt
@@ -10,7 +10,6 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.RelativeLayout
 import android.widget.TextView
-import android.widget.Toast
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.DefaultItemAnimator
@@ -18,19 +17,13 @@ import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.bumptech.glide.Glide
-import com.example.movie.BuildConfig
-import com.example.movie.DetailActivity
+import com.example.movie.view.DetailActivity
 import com.example.movie.R
 import com.example.movie.adapter.MoviesAdapter
-import com.example.movie.api.RetrofitService
-import com.example.movie.database.MovieDao
-import com.example.movie.database.MovieDatabase
 import com.example.movie.model.Movie
 import com.example.movie.model.Singleton
 import com.example.movie.view_model.MovieListViewModel
 import com.example.movie.view_model.ViewModelProviderFactory
-import kotlinx.coroutines.*
-import kotlin.coroutines.CoroutineContext
 
 /**
  * A simple [Fragment] subclass.
