+++ app/src/main/java/com/example/movie/view_model/MovieListViewModel.kt
@@ -8,7 +8,6 @@ import com.example.movie.api.RetrofitService
 import com.example.movie.database.MovieDao
 import com.example.movie.database.MovieDatabase
 import com.example.movie.model.Movie
-import com.google.gson.JsonObject
 import kotlinx.coroutines.*
 import kotlin.coroutines.CoroutineContext
 
@@ -21,7 +20,6 @@ class MovieListViewModel(
     private var sessionId: String? = ""
 
     val liveData = MutableLiveData<List<Movie>>()
-    val liveDataLike = MutableLiveData<List<Movie>>()
 
     init {
         movieDao = MovieDatabase.getDatabase(context = context).movieDao()
@@ -33,89 +31,6 @@ class MovieListViewModel(
     override fun onCleared() {
         super.onCleared()
         job.cancel()
-    }
-
-    fun getMovieLike() {
-
-        launch {
-
-            val likesOffline = movieDao.getLikedOffline(11)
-
-            for (i in likesOffline) {
-                val body = JsonObject().apply {
-                    addProperty("media_type", "movie")
-                    addProperty("media_id", i)
-                    addProperty("favorite", true)
-                }
-                try {
-                    RetrofitService.getPostApi()
-                        .rateCoroutine(
-                            accountId,
-                            BuildConfig.THE_MOVIE_DB_API_TOKEN,
-                            sessionId,
-                            body
-                        )
-                } catch (e: Exception) {
-
-                }
-            }
-
-
-            val unLikesOffline = movieDao.getLikedOffline(10)
-
-            for (i in unLikesOffline) {
-                val body = JsonObject().apply {
-                    addProperty("media_type", "movie")
-                    addProperty("media_id", i)
-                    addProperty("favorite", false)
-                }
-                try {
-                    RetrofitService.getPostApi()
-                        .rateCoroutine(
-                            accountId,
-                            BuildConfig.THE_MOVIE_DB_API_TOKEN,
-                            sessionId,
-                            body
-                        )
-                } catch (e: Exception) {
-                }
-            }
-
-            val unLikeMoviesOffline = movieDao.getUnLikedOffline()
-            val newArray: ArrayList<Movie>? = null
-            for (movie in unLikeMoviesOffline) {
-                movie.liked = 0
-                newArray?.add(movie)
-            }
-             newArray?.let { movieDao.insertAll(it) }
-
-            val list = withContext(Dispatchers.IO) {
-                try {
-                    val response = RetrofitService.getPostApi().getFavouriteMoviesCoroutine(
-                        accountId,
-                        BuildConfig.THE_MOVIE_DB_API_TOKEN,
-                        sessionId
-                    )
-                    if (response.isSuccessful) {
-                        val result = response.body()?.results
-                        if (result != null) {
-                            for (m in result) {
-                                m.liked = 1
-                            }
-                        }
-                        if (!result.isNullOrEmpty()) {
-                            movieDao.insertAll(result)
-                        }
-                        result
-                    } else {
-                        movieDao.getAllLiked()
-                    }
-                } catch (e: Exception) {
-                    movieDao.getAllLiked()
-                }
-            }
-            liveDataLike.postValue(list)
-        }
     }
 
     fun getMoviesList() {
