+++ app/src/main/res/layout/activity_main.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".MainActivity">
+        tools:context=".view.MainActivity">
 
         <include layout="@layout/content_main" />
