+++ app/src/main/res/layout/activity_registration.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegistrationActivity"
+    tools:context=".view.RegistrationActivity"
     android:background = "#ffffff"
     android:orientation="vertical">
     <View
