+++ app/src/main/res/layout/content_main.xml
@@ -6,7 +6,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:background="@color/viewBg"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.example.movie.MainActivity"
+    tools:context="com.example.movie.view.MainActivity"
     tools:showIn="@layout/activity_main"
     >
